+++ Source/Vertices/Public/VerticesApiOperations.h
@@ -144,7 +144,7 @@ namespace algorand {
             TOptional<FString> senderAddress;
             TOptional<FString> receiverAddress;
             TOptional<uint64_t> asset_id;
-            TOptional<uint64_t> amount;
+            TOptional<double> amount;
             TOptional<FString> notes;
         };
